requirements.txt
numpy~=1.26.0
opencv-python~=4.8.0.76
flask~=2.3.3
Pillow~=10.0.1
ultralytics~=8.0.184